🔒 Full CRUD Roles Immune Master Account Roles Page More Security when changing roles (send help I have eye strain lol)

+++ app/api/admin/users/route.js
@@ -1,4 +1,3 @@
-import { handleAdmin } from "@/lib/utils"
 import { handle } from "@/lib/utils"
 import { encryptData, decryptData } from "@/lib/utils"
 import prisma from "@/prisma/prisma-client"
@@ -7,7 +6,7 @@ const secret = process.env.SECRET
 
 export const GET = async (req) => {
   return await handle(req,["READ_USERS"],async () => {
-    const users = await prisma.user.findMany()
+    const users = await prisma.user.findMany({include:{role:true}})
     users.forEach((user) => {
       user.password = decryptData(user.password,secret)
     })
@@ -16,12 +15,20 @@ export const GET = async (req) => {
 }
 
 export const POST = async (req) => {
-  return await handleAdmin(req, async () => {
+  return await handle(req,["WRITE_USERS"], async () => {
     const body = await req.json()
     const {name,email,password,roleId} = body
     if(!name || !email || !password || !roleId){
       console.log(name,email,password,roleId)
       return Response.json("Bad Request", {status: 400})
+    }
+    const roleIsMaster = await prisma.role.findFirst({
+      where:{
+        id: roleId
+      }
+    })
+    if(roleIsMaster.name == "MASTER"){
+      return Response.json("Rejection", {status: 401})
     }
     const encyptedPassword = encryptData(password,secret)
     const newUser = await prisma.user.create({
@@ -30,44 +37,66 @@ export const POST = async (req) => {
         email,
         password: encyptedPassword,
         roleId
-      },
+      },include:{role:true}
     })
     return Response.json({...newUser,password}, {status: 201})
   })
 }
 
 export const PUT = async (req) => {
-  return await handleAdmin(req, async () => {
+  return await handle(req,["WRITE_USERS"], async ({tokenRes}) => {
     const body = await req.json()
     const {id,name,email,password,roleId} = body
     if(!id || !name || !email || !password || !roleId){
       return Response.json("Bad Request", {status: 400})
     }
+    if(id == tokenRes.body.user.id){
+      return Response.json("You can't edit your own account, use Another.", {status: 401})
+    }
+    const user = await prisma.user.findFirst({where:{id},include:{role:true}})
+    if(user.role.name == "MASTER"){
+      return Response.json("Rejection", {status: 401})
+    }
+    const roleIsMaster = await prisma.role.findFirst({
+      where:{
+        id: roleId
+      }
+    })
+    if(roleIsMaster.name == "MASTER"){
+      return Response.json("Rejection", {status: 401})
+    }
     const encyptedPassword = encryptData(password,secret)
     const updatedUser = await prisma.user.update({
-      where: { id: id },
+      where: {id},
       data: {
         name,
         email,
         password: encyptedPassword,
         roleId
-      },
+      },include:{role:true}
     })
     return Response.json({...updatedUser,password}, {status: 200})
   })
 }
 
 export const DELETE = async (req) => {
-  return await handleAdmin(req, async () => {
+  return await handle(req,["WRITE_USERS"], async ({tokenRes}) => {
     const { id } = await req.json()
     if(!id){
       return Response.json("Bad Request", {status: 400})
+    }
+    if(id == tokenRes.body.user.id){
+      return Response.json("You can't delete your own account, use Another.", {status: 401})
+    }
+    const FoundUser = await prisma.user.findFirst({where:{id},include:{role:true}})
+    if(FoundUser.role.name == "MASTER"){
+      return Response.json("Rejection", {status: 401})
     }
     await prisma.application.deleteMany({
       where: { userId: id },
     })
     const deletedUser = await prisma.user.delete({
-      where: { id: id },
+      where: {id },
     })
     return Response.json(deletedUser, {status: 200})
   })
